+++ modules/angular2/src/render/xhr.ts
