+++ modules/angular2/src/render/xhr_impl.ts
