+++ modules/angular2/src/transform/common/xhr_impl.dart
@@ -1,7 +1,7 @@
 library angular2.transform.template_compiler.xhr_impl;
 
 import 'dart:async';
-import 'package:angular2/src/services/xhr.dart' show XHR;
+import 'package:angular2/src/render/xhr.dart' show XHR;
 import 'package:angular2/src/transform/common/asset_reader.dart';
 import 'package:angular2/src/transform/common/logging.dart';
 import 'package:barback/barback.dart';
